refactor(typeagent-core): dedupe agent-name resolution
Extract resolveAgentName into a dependency-free sandbox/agentRef module.
inMemorySandboxManager's default loader now uses it instead of a private
deriveAgentName subset (which mishandled packages/agents/<name> paths);
repoAgentLoader re-exports it so its API and tests are unchanged.
